back/requirements.txt
Django==5.2
django-extensions==4.1
django-ninja==1.4.1
pytest==8.3.5
pytest-django==4.11.1
requests==2.32.3
factory-boy==3.3.3
django-ninja-jwt==5.3.7